Future weather forecast
Removed the old icons and days in HTML and replaced it with new content from Week 8 Lesson 3 and 4.

I also added/updated weather-temperature so that the image icon and the temperatures (plus c and f) moved to the left and then the image aligns with the search button.

In JS I learnt how to repeat the code using days so that the code wasn't physically copied multiple times but would still be displayed multiple times.