Update JS - API working
Store response from the api in a variable called Forecast. It looks through this variable and it shows the next upcoming days including the current day.

Using the object and index for each function to format the day of the time stamp using a function I created. Includes the icon name and rounding the temperature for max and min temperatures.

Set up the index so that it only shows 6 days.

Week 8 Lesson 6.